requirements.txt
streamlit
pandas
matplotlib
wordcloud
plotly
requests
beautifulsoup4
Pillow